+++ release_notes.md
@@ -19,6 +19,8 @@ Changed behaviour:
 - SPECT ray tracer projectors now normalize emission probability based on a pre-determined superellipse
   - Defaults to the superellipse corresponding to FOV boundary
   - Traced rays are clipped to the inside of the superellipse
+  - Currently only ```options.ellipsePower``` values of 2 (elliptic cylinder, an ellipse in the transaxial
+    plane) and ```Inf``` (box, the default) are supported; other superellipse powers raise an error
 - Memory optimization for SPECT
   - These arrays can now be input as size nRowsD * nColsD * nHeads (one image for each detector head):
     - FP mask